Address PR comments: Move SharedTestOutputHelper to its own file
