+++ test/dotnet-new3.UnitTests/SharedHomeDirectory.cs
@@ -75,29 +75,5 @@ namespace dotnet_new3.UnitTests
                 .And
                 .NotHaveStdErr();
         }
-
-        /// <summary>
-        /// This is so we can pass ITestOutputHelper to TestCommand constructor
-        /// when calling from SharedHomeDirectory
-        /// </summary>
-        class SharedTestOutputHelper : ITestOutputHelper
-        {
-            private readonly IMessageSink sink;
-
-            public SharedTestOutputHelper(IMessageSink sink)
-            {
-                this.sink = sink;
-            }
-
-            public void WriteLine(string message)
-            {
-                sink.OnMessage(new DiagnosticMessage(message));
-            }
-
-            public void WriteLine(string format, params object[] args)
-            {
-                sink.OnMessage(new DiagnosticMessage(format, args));
-            }
-        }
     }
 }
